Fix country code

+++ programs/nosana-nodes/src/instructions/register.rs
@@ -22,7 +22,7 @@ impl<'info> Register<'info> {
     pub fn handler(
         &mut self,
         architecture_type: u8,
-        country_code: u8,
+        country_code: u16,
         cpu: u16,
         gpu: u16,
         memory: u16,
